Fix connection name missing while executing the operation
