+++ tests/src/test/java/org/ballerina/test/TestArrayConnector.java
@@ -46,6 +46,7 @@ public class TestArrayConnector {
         // Create a context for connection initialization
         TestMessageContext initContext = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
+                .isConnection(true)
                 .addParameter("apiUrl", "string", "http://test.api.com")
                 .addParameter("connectionType", "string", "ARRAYPROJECT_ARRAYCLIENT")
                 .build();
@@ -296,85 +297,5 @@ public class TestArrayConnector {
 
         String result = ((DefaultConnectorResponse) context.getVariable("result")).getPayload().toString();
         Assert.assertEquals(result, "[]", "Should return empty JSON array");
-    }
-
-    /**
-     * Helper class to build test MessageContext for connector tests.
-     */
-    static class ConnectorContextBuilder {
-        private int paramCount = 0;
-        private String connectionName;
-        private String methodName;
-        private String returnType;
-        private HashMap<String, Object> properties = new HashMap<>();
-        private HashMap<Object, Object> templateParams = new HashMap<>();
-
-        static ConnectorContextBuilder connectorContext() {
-            return new ConnectorContextBuilder();
-        }
-
-        public ConnectorContextBuilder connectionName(String name) {
-            this.connectionName = name;
-            return this;
-        }
-
-        public ConnectorContextBuilder methodName(String name) {
-            this.methodName = name;
-            return this;
-        }
-
-        public ConnectorContextBuilder objectTypeName(String name) {
-            properties.put("objectTypeName", name);
-            return this;
-        }
-
-        public ConnectorContextBuilder returnType(String type) {
-            this.returnType = type;
-            return this;
-        }
-
-        public ConnectorContextBuilder addParameter(String name, String type, String value) {
-            properties.put("param" + paramCount, name);
-            properties.put("paramType" + paramCount, type);
-            templateParams.put(name, value);
-            paramCount++;
-            return this;
-        }
-
-        public ConnectorContextBuilder addProperty(String key, Object value) {
-            properties.put(key, value);
-            return this;
-        }
-
-        public TestMessageContext build() {
-            TestMessageContext context = new TestMessageContext();
-
-            // Set connector-specific properties
-            if (connectionName != null) {
-                properties.put("connectionName", connectionName);
-            }
-            if (methodName != null) {
-                properties.put("paramFunctionName", methodName);  // Use correct constant
-            }
-            if (returnType != null) {
-                properties.put("returnType", returnType);
-            }
-            properties.put("paramSize", paramCount);
-
-            for (var entry : properties.entrySet()) {
-                context.setProperty(entry.getKey(), entry.getValue());
-            }
-
-            // Set up template context for parameter lookup
-            Stack<TemplateContext> stack = new Stack<>();
-            TemplateContext templateContext = new TemplateContext("testConnectorFunc", new ArrayList<>());
-            templateParams.put("responseVariable", "result");
-            templateParams.put("connectionName", connectionName != null ? connectionName : CONNECTION_NAME);
-            templateContext.setMappedValues(templateParams);
-            stack.push(templateContext);
-            context.setProperty("_SYNAPSE_FUNCTION_STACK", stack);
-
-            return context;
-        }
     }
 }
