+++ tests/src/test/java/org/ballerina/test/TestJsonType.java
@@ -35,8 +35,9 @@ public class TestJsonType {
         BalConnectorConfig config = new BalConnectorConfig(moduleInfo);
 
         // Create a context for connection initialization
-        TestMessageContext initContext = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext initContext = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
+                .isConnection(true)
                 .objectTypeName("JsonClient")
                 .addParameter("serviceUrl", "string", "http://test.api.com")
                 .addParameter("connectionType", "string", "JSONPROJECT_JSONCLIENT")
@@ -56,7 +57,7 @@ public class TestJsonType {
     public void testGetJsonAsString() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("getJsonAsString")
                 .returnType("string")
@@ -79,7 +80,7 @@ public class TestJsonType {
     public void testCreateJsonObject() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("createJsonObject")
                 .returnType("json")
@@ -107,7 +108,7 @@ public class TestJsonType {
     public void testMergeJson() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("mergeJson")
                 .returnType("json")
@@ -134,7 +135,7 @@ public class TestJsonType {
     public void testGetArrayLength() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("getArrayLength")
                 .returnType("int")
@@ -157,7 +158,7 @@ public class TestJsonType {
     public void testTransformJson() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("transformJson")
                 .returnType("json")
@@ -185,7 +186,7 @@ public class TestJsonType {
     public void testGetNestedValue() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("getNestedValue")
                 .returnType("string")
@@ -208,7 +209,7 @@ public class TestJsonType {
     public void testSumJsonArray() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("sumJsonArray")
                 .returnType("int")
@@ -232,7 +233,7 @@ public class TestJsonType {
         BalConnectorFunction connector = new BalConnectorFunction();
 
         // Test with existing key
-        TestMessageContext contextExists = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext contextExists = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("hasKey")
                 .returnType("boolean")
@@ -254,7 +255,7 @@ public class TestJsonType {
         Assert.assertEquals(resultExists, "true", "Should return true for existing key");
 
         // Test with non-existing key
-        TestMessageContext contextNotExists = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext contextNotExists = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("hasKey")
                 .returnType("boolean")
@@ -280,7 +281,7 @@ public class TestJsonType {
     public void testGetArrayElement() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("getArrayElement")
                 .returnType("json")
@@ -306,7 +307,7 @@ public class TestJsonType {
     public void testCreateNestedJson() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("createNestedJson")
                 .returnType("json")
@@ -338,7 +339,7 @@ public class TestJsonType {
         BalConnectorFunction connector = new BalConnectorFunction();
 
         String jsonArray = "[{\"id\":1,\"status\":\"active\"},{\"id\":2,\"status\":\"inactive\"},{\"id\":3,\"status\":\"active\"}]";
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("filterJsonArray")
                 .returnType("json")
@@ -366,7 +367,7 @@ public class TestJsonType {
     public void testCountKeys() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("countKeys")
                 .returnType("int")
