+++ tests/src/test/java/org/ballerina/test/TestRecordType.java
@@ -35,8 +35,9 @@ public class TestRecordType {
         BalConnectorConfig config = new BalConnectorConfig(moduleInfo);
 
         // Create a context for connection initialization
-        TestMessageContext initContext = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext initContext = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
+                .isConnection(true)
                 .objectTypeName("RecordClient")
                 .addParameter("serviceUrl", "string", "http://test.api.com")
                 .addParameter("connectionType", "string", "RECORDPROJECT_RECORDCLIENT")
@@ -56,7 +57,7 @@ public class TestRecordType {
     public void testSimpleRecord() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("simpleRecordFunction")
                 .returnType("string")
@@ -80,7 +81,7 @@ public class TestRecordType {
     public void testNestedRecord() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("getUserSummary")
                 .returnType("string")
@@ -104,7 +105,7 @@ public class TestRecordType {
     public void testOptionalFieldsAndArrays() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("summarizeOrder")
                 .returnType("string")
@@ -128,7 +129,7 @@ public class TestRecordType {
     public void testNestedArraysSum() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("computeCatalogTotal")
                 .returnType("float")
@@ -154,7 +155,7 @@ public class TestRecordType {
         BalConnectorFunction connector = new BalConnectorFunction();
 
         // Present
-        TestMessageContext ctxPresent = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext ctxPresent = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("formatProfile")
                 .returnType("string")
@@ -174,7 +175,7 @@ public class TestRecordType {
         Assert.assertEquals(resPresent, "u1:true:2", "Profile present should reflect emailOptIn and tag count");
 
         // Absent
-        TestMessageContext ctxAbsent = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext ctxAbsent = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("formatProfile")
                 .returnType("string")
@@ -198,7 +199,7 @@ public class TestRecordType {
     public void testGetUppercasedPerson() throws Exception {
         BalConnectorFunction connector = new BalConnectorFunction();
 
-        TestMessageContext context = TestArrayConnector.ConnectorContextBuilder.connectorContext()
+        TestMessageContext context = ConnectorContextBuilder.connectorContext()
                 .connectionName(CONNECTION_NAME)
                 .methodName("getUppercasedPerson")
                 .returnType("record")
